Fixed "remove killer again" bug.
A bug existed where the program would crash if the user received a killer they previously removed from the active killer pool and tried to remove them again. The list being pulled from was the master list rather than the pool list, therefore if the user removed the presented option from the pool, the program would still be able to receive the same choice again rather than not, and the program couldn't find in the pool, the option the program presented again.